Added getEmptyDeck method
Added a method that creates an empty deck. Will probably move to hand typing all the associated suits and values. This is so I don't have to keep retyping or reworking this method when moving or adding Suits and Values or different Card types. As well as when I create a different collection for cards such as the Shoe, I don't have to retype the for lists to generate a deck of cards.